+++ tests/conftest.py
@@ -1,35 +1,16 @@
 import os
-from unittest.mock import Mock, MagicMock
+from unittest.mock import MagicMock
 
 import pytest
+from aws_xray_sdk import core as xray_core
 
 from coworks.cws.fixture import local_server_factory as factory
-from tests.mockup import email_mock, smtp_mock, boto3_mock
+from tests.mockup import email_mock, smtp_mock
 
 
 @pytest.fixture
 def example_dir():
     return os.getenv('EXAMPLE_DIR')
-
-
-@pytest.fixture
-def zipfile_mock(request):
-    import zipfile
-
-    file = MagicMock()
-    file.__enter__ = Mock(return_value=file)
-    file.write = Mock()
-    zip_mock = Mock(return_value=file)
-    zip_mock.file = file
-
-    save_zip_file = zipfile.ZipFile
-    zipfile.ZipFile = zip_mock
-
-    def end():
-        zipfile.ZipFile = save_zip_file
-
-    request.addfinalizer(end)
-    return zip_mock
 
 
 @pytest.fixture
@@ -42,9 +23,18 @@ def smtp_mock_fixture():
     yield smtp_mock
 
 
+s3session_mock = MagicMock()
+
+
 @pytest.fixture
-def boto3_mock_fixture():
-    yield boto3_mock
+def s3_session():
+    class S3Session:
+        mock = s3session_mock
+
+        def __new__(cls, *args, **kwargs):
+            return s3session_mock
+
+    yield S3Session
 
 
 local_server_factory = factory
@@ -54,3 +44,7 @@ def pytest_sessionstart():
     example_dir = os.getenv('EXAMPLE_DIR')
     if not os.path.exists(example_dir):
         raise pytest.UsageError(f"Undefined example folder: {example_dir} (value defined in pytest.ini).")
+
+    # mock aws
+    xray_core.recorder.capture = lambda _: lambda y: y
+    xray_core.recorder.current_subsegment = lambda: MagicMock()