+++ tests/example/quickstart1.py
@@ -5,7 +5,7 @@ from coworks.cws.runner import CwsRunner
 class SimpleMicroService(TechMicroService):
 
     def get(self):
-        return f"Simple microservice ready.\n"
+        return "Simple microservice ready.\n"
 
 
 app = SimpleMicroService()