+++ tests/example/quickstart2.py
@@ -1,25 +1,17 @@
 from coworks import TechMicroService
-from coworks.config import Config
+from coworks.cws.deployer import CwsTerraformDeployer
 from coworks.cws.runner import CwsRunner
-from coworks.cws.writer import CwsTerraformWriter
 
 
 class SimpleMicroService(TechMicroService):
 
-    def get(self):
-        return f"Simple microservice ready.\n"
-
     def auth(self, auth_request):
-        return auth_request.token == "token"
+        return True
+
+    def get(self):
+        return "Simple microservice ready.\n"
 
 
-CONFIG = Config(
-    workspace="dev"
-)
-
-app = SimpleMicroService(name='test', configs=[CONFIG])
+app = SimpleMicroService()
 CwsRunner(app)
-CwsTerraformWriter(app, name='export')
-
-if __name__ == '__main__':
-    app.execute('run', project_dir='.', workspace='dev', module='quickstart')
+CwsTerraformDeployer(app, name='deploy')