+++ tests/example/test_example.py
@@ -13,9 +13,10 @@ from .example import TechMS
 
 
 class TestClass:
+
     def test_simple_example(self, local_server_factory):
         local_server = local_server_factory(TechMS())
-        response = local_server.make_call(requests.get, '/')
+        response = local_server.make_call(requests.get, '/', timeout=500)
         assert response.status_code == 200
         assert response.text == "Simple microservice for test.\n"
         response = local_server.make_call(requests.get, '/', params={"usage": "demo"})
@@ -69,7 +70,7 @@ class TestClass:
         assert response.status_code == 200
         assert response.text == "Simple microservice for demo.\n"
 
-    def test_cmd_wront_project_dir(self, example_dir, capsys):
+    def test_cmd_wrong_project_dir(self, example_dir, capsys):
         with pytest.raises(SystemExit) as pytest_wrapped_e:
             client(prog_name='cws', args=['-p', 'tests/example', 'info'], obj={})
         assert pytest_wrapped_e.type == SystemExit