+++ tests/src/cws/test_cws.py
@@ -79,6 +79,6 @@ class TestClass:
     def test_export(self, example_dir):
         with pytest.raises(SystemExit) as pytest_wrapped_e:
             client(prog_name='cws',
-                   args=['-p', example_dir, '-m', 'example', '-s', 'tech_app', 'terraform', '--config', None], obj={})
+                   args=['-p', example_dir, '-m', 'example', '-s', 'tech_app', 'export'], obj={})
         assert pytest_wrapped_e.type == SystemExit
         assert pytest_wrapped_e.value.code == 0